+++ src/vector.cpp
@@ -26,7 +26,7 @@ Vector3 operator/(const Vector3 &v1, const Vector3 &v2) {
     return Vector3(v1.x / v2.x, v1.y / v2.y, v1.z / v2.z);
 }
 
-Vector3 operator/(const Vector3 &v1, double s) {
+Vector3 operator/(const Vector3 &v1, float s) {
     return Vector3(v1.x / s, v1.y / s, v1.z / s);
 }
 std::ostream &operator<<(std::ostream &os, const Vector3 &v) {
